pyproject.toml
[project]
name = "external-context-engine"
version = "0.1.0"
description = "An external context engine for LLMs using the Elysia framework."
authors = [
    {name = "Rob Balch II", email = "robert.balch.ii@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elysia-ai",
    "python-dotenv",
    "pyyaml",
    "neo4j",
    "ollama",
    "pydantic",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]


